resources/solutions/ai-specialist/model_test_app/object-detection-batch-model-api/ui/triton-pytorchscript/requirements.txt
gradio
opencv-python
numpy
Pillow
requests
tritonclient[http]